refactor(ai): extract AIChatFeed and modularize sidebar subcomponents
